% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasification.R
\name{rvClassifier}
\alias{rvClassifier}
\title{R-vine Classification}
\usage{
rvClassifier(data, n, prob, vine, copula, margin, tree, k, nRuns)
}
\arguments{
\item{data}{a list of four elements: in the first position, it has the data matrices (dataTrain)
for each class; in the second position, the test matrix (dataTest); in the third position,
a list (N_train) with the sample size of each training matrix; and in the fourth position, 
a list (N_test) with the sample size for each class in the test matrix. These matrices 
and vectors must be organized by class from class 1 to \code{nc}.}

\item{n}{the number of variables (or classification features), i.e., the dimension of the problem.}

\item{prob}{a vector with the probability of each class.}

\item{vine}{a vector that specifies the type of vine copula per class.}

\item{copula}{a vector that specifies the copula families to be fitted in the vine of each class.}

\item{margin}{a vector that specifies the marginal family of each class.}

\item{tree}{a vector that specifies the number of trees in the vine of each class.}

\item{k}{the number of partitions of the k-fold crossvalidation.}

\item{nRuns}{the number of crossvalidation repetitions.}
}
\value{
\code{predict}, \code{validateModel} A list with two elements: The first, \code{predict}, is 
        the confusion matrix with the prediction results in each fold. The second, \code{validateModel}, is 
        mean-confusion matrix, which is computed from confusion matrices in code{predict}.
}
\description{
Lunch the R-vine classification workflow
}

\examples{
# Create a sample with 2 classes.

nc = 2

class1 <- matrix(runif(100,0,1), ncol= 10, nrow=10)
class2 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
classes <- list(class1,class2)

test1 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
test2 <- matrix(runif(300,0,1), ncol= 10, nrow=30)
test <- rbind(test1,test2)

N_classes <- list(10,20)
N_test <- list(20,30)

data <- list(classes,test,N_classes, N_test)
prob <- c(0.33, 0.66)

k <- 5
nRuns <- 1
n <- dim(data[[1]][[1]])[[2]]
      
# Using only normal copula for each class.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL))

m <- rvClassifier(data, prob, n, vine, copula, margin, tree, k, nRuns)

# Create a sample with 5 classes.
nc = 5 

# Using diferent type of copula per class.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL, COPULA_CLAYTON),c(COPULA_CLAYTON),c(COPULA_NORMAL,COPULA_CLAYTON),c(COPULA_NORMAL),c(COPULA_NORMAL))
}
