% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasification.R
\name{rvEvaluate}
\alias{rvEvaluate}
\title{Model Evaluation}
\usage{
rvEvaluate(prec, N_test)
}
\arguments{
\item{prec}{a list with the class predicted for each vector of the test set.}

\item{N_test}{a list with the sample size of each test matrix (one per class).}
}
\value{
\code{confusionM} a confusion matrix.
}
\description{
Calculate a confusion matrix in order to describe the performance of the vine classifier
}

\examples{
# Create a sample with 2 classes.

nc = 2

class1 <- matrix(runif(100,0,1), ncol= 10, nrow=10)
class2 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
classes <- list(class1,class2)

test1 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
test2 <- matrix(runif(300,0,1), ncol= 10, nrow=30)
test <- rbind(test1,test2)

N_classes <- list(10,20)
N_test <- list(20,30)

data <- list(classes,test,N_classes, N_test)
prob <- c(0.33, 0.66)

n <- dim(data[[1]][[1]])[[2]]
      
# Using only normal copula for each class.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL))

models <- rvSpecify(n, vine, margin, tree)
trainModel <- rvLearn(data[[1]], models, copula)
precModel <- rvPredict(trainModel,data[[2]], data[[4]], prob)
evalM <- rvEvaluate(precModel, data[[4]])
}


