% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasification.R
\name{rvLearn}
\alias{rvLearn}
\title{Model Learning}
\usage{
rvLearn(data, models, copula)
}
\arguments{
\item{data}{a list of matrices with the training set of each class.}

\item{models}{a list of vectors that specifies the type of vine, of marginal, number of trees that you want 
to learn for each class. The value of models is the value of the rvSpecified function.}

\item{copula}{a vector where specifying the copulas to be adjusted for each class.}
}
\value{
\code{mvdvList} a vector that specifies the copula families of each class to be fitted in the 
        corresponding vine.
}
\description{
Learn a vine for each class
}

\examples{
# Create a sample with 2 classes.

nc = 2

class1 <- matrix(runif(100,0,1), ncol= 10, nrow=10)
class2 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
classes <- list(class1,class2)

test1 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
test2 <- matrix(runif(300,0,1), ncol= 10, nrow=30)
test <- rbind(test1,test2)

N_classes <- list(10,20)
N_test <- list(20,30)

data <- list(classes,test,N_classes, N_test)
prob <- c(0.33, 0.66)

n <- dim(data[[1]][[1]])[[2]]
      
# Using only normal copula for each class.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL))

models <- rvSpecified(n, vine, margin, tree)
trainModel <- rvLearn(data[[1]], models, copula)
}


