% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasification.R
\name{rvPredict}
\alias{rvPredict}
\title{Model Prediction}
\usage{
rvPredict(vineModel, testSet, N_test, prob)
}
\arguments{
\item{vineModel}{a list with the models learned for each class by the function \code{rvLearn}}

\item{testSet}{the test matrix that contains the test set of each class. These matrices are concateneted
by the row axis (i.e., one matrix after the other).}

\item{N_test}{a list with the sample size of each test matrix (one per class).}

\item{prob}{a vector with the probability of each class.}
}
\value{
\code{pred} a list with the class predicted for each row vector from the test matrix.
}
\description{
Predict the class of each vector of features
}


\examples{
# Create a sample with 2 classes.

nc = 2

class1 <- matrix(runif(100,0,1), ncol= 10, nrow=10)
class2 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
classes <- list(class1,class2)

test1 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
test2 <- matrix(runif(300,0,1), ncol= 10, nrow=30)
test <- rbind(test1,test2) 

N_classes <- list(10,20)
N_test <- list(20,30)

data <- list(classes,test,N_classes, N_test)
prob <- c(0.33, 0.66)

n <- dim(data[[1]][[1]])[[2]]
      
# Using only normal copula for each class.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL))

models <- rvSpecify(n, vine, margin, tree)
trainModel <- rvLearn(data[[1]], models, copula)
precModel <- rvPredict(trainModel,data[[2]], data[[4]], prob)
}

