% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasification.R
\name{rvSpecified}
\alias{rvSpecified}
\title{Model Specification}
\usage{
rvSpecified(n, vine, margin, tree)
}
\arguments{

\item{n}{the number of variables.}

\item{vine}{a vector that specifies the type of vine copula per class.}

\item{margin}{a vector that specifies the marginal family of each class.}

\item{tree}{a vector that specifies the number of trees in the vine of each class.}
}
\value{
\code{vector} a list with the models (a model = vine + margins) that have been specified for each class.
}
\description{
Specifies the configuration of the vine
}

\examples{
# Create a sample with 2 classes.

n <- 10 #Number of variables.
nc <- 2
      
# Using only normal copula for each class and Rvine.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL))

models <- rvSpecified(n, vine, margin, tree)

# Using diferent types of copula and diferent levels of tree.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(1,2)
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL, COPULA_CLAYTON), c(COPULA_GUMBEL, COPULA_NORMAL), c(COPULA_CLAYTON), c(COPULA_NORMAL))

models <- rvSpecified(n, vine, margin, tree)

# Using Dvine and diferent types of margin.
vine <- c(rep(VINE_DVINE,nc))
margin <- c(MARGIN_NORMAL,MARGIN_GAMMA)
tree <- c(1,2)
copula <- list(c(COPULA_GUMBEL),c(COPULA_NORMAL, COPULA_CLAYTON))

models <- rvSpecified(n, vine, margin, tree)
}

