% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasification.R
\name{rvValidate}
\alias{rvValidate}
\title{Model Validation}
\usage{
rvValidate(data, n, N_train, k, nRuns, models, copula)
}
\arguments{
\item{data}{a list of matrices where each matrix is the training set of each class.}

\item{n}{the number of variables.}

\item{N_train}{a list with the sample size of each training matrix (one per class).}

\item{k}{the number of partitions of the k-fold crossvalidation.}

\item{nRuns}{the number of crossvalidation repetitions.}

\item{models}{a list of vectors that specifies the type of vines, types of margins, and the number of trees of 
the vine of each class. The value of models is the value of the rvSpecified function.}

\item{copula}{a vector that specifies the copula families to be fitted in the vine of each class.}
}
\value{
\code{confusionMatrix} a confusion matrix, which is the mean of the \code{nRuns} confusion 
        matrices obtained in each crossvalidation's fold.
}
\description{
Validate the model through the crossvalidation method
}

\examples{
# Create a sample with 2 classes.
nc = 2

class1 <- matrix(runif(100,0,1), ncol= 10, nrow=10)
class2 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
classes <- list(class1,class2)

test1 <- matrix(runif(200,0,1), ncol= 10, nrow=20)
test2 <- matrix(runif(300,0,1), ncol= 10, nrow=30)
test <- rbind(test1,test2)

N_classes <- list(10,20)
N_test <- list(20,30)

data <- list(classes,test,N_classes, N_test)
prob <- c(0.33, 0.66)

k <- 5
nRuns <- 5
n <- dim(data[[1]][[1]])[[2]]
      
# Using only normal copula for each class.
vine <- c(rep(VINE_RVINE,nc))
margin <- c(rep(MARGIN_NORMAL,nc))
tree <- c(rep(1,nc))
copula <- list(c(COPULA_NORMAL),c(COPULA_NORMAL))

models <- rvSpecify(n, vine, margin, tree)
validatM <- rvValidate(data[[1]], n, data[[3]], k, nRuns, models, copula)
}


